# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HermesX API
  description: Agent-first Runtime Control Plane for governed, multi-tenant AI automation.
  version: 2.4.0-dev
  contact:
    name: HermesX Team
    url: https://github.com/Colin4k1024/hermesx
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.hermesx.dev
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /v1/chat/completions:
    post:
      tags: [Agent]
      summary: Create chat completion (OpenAI-compatible)
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/agent/chat:
    post:
      tags: [Agent]
      summary: Agent chat with tool execution
      operationId: agentChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentChatRequest'
      responses:
        '200':
          description: Agent chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentChatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/sessions:
    get:
      tags: [Sessions]
      summary: List user sessions
      operationId: listSessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Sessions]
      summary: Create a new session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/memories:
    get:
      tags: [Memory]
      summary: List memories
      operationId: listMemories
      parameters:
        - name: X-Hermes-User-Id
          in: header
          schema:
            type: string
          description: User ID to query (admin only)
      responses:
        '200':
          description: List of memories
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id:
                    type: string
                  user_id:
                    type: string
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemoryEntry'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/skills:
    get:
      tags: [Skills]
      summary: List available skills
      operationId: listSkills
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/gdpr/export:
    get:
      tags: [GDPR]
      summary: Export user data
      operationId: exportGDPRData
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/gdpr/data:
    delete:
      tags: [GDPR]
      summary: Delete user data
      operationId: deleteGDPRData
      responses:
        '200':
          description: Data deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/gdpr/restore:
    post:
      tags: [GDPR]
      summary: Restore deleted data
      operationId: restoreGDPRData
      responses:
        '200':
          description: Data restored
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/gdpr/status:
    get:
      tags: [GDPR]
      summary: Get GDPR deletion status
      operationId: getGDPRStatus
      responses:
        '200':
          description: Deletion status
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key in format hk_...
  schemas:
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
        temperature:
          type: number
        max_tokens:
          type: integer
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
    AgentChatRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
        tools:
          type: array
          items:
            type: object
        stream:
          type: boolean
    AgentChatResponse:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        tool_calls:
          type: array
          items:
            type: object
...
    Session:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        user_id:
          type: string
        model:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
    SessionCreateRequest:
      type: object
      properties:
        model:
          type: string
        user_id:
          type: string
...
    MemoryEntry:
      type: object
      properties:
        key:
          type: string
        content:
          type: string
    Skill:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: authorization required
...